[MODIFY] api 이름 변경 및 라우트 정리
