+++ src/api/delete/deleteComments.api.js
@@ -1,4 +1,4 @@
-import client from '../config/api.config';
+import client from '../../config/api.config';
 
 const deleteComments = async (id, commentId) => {
   const token = JSON.parse(localStorage.getItem('user')).token;
