+++ src/api/get/getCheckToken.api.js
@@ -1,4 +1,4 @@
-import client from '../config/api.config';
+import client from '../../config/api.config';
 
 const getCheckToken = (token) => {
   return client.get(`/user/checktoken`, { ...client.AuthType(token) });
