+++ src/api/get/getFollowingList.api.js
@@ -1,6 +1,6 @@
 import { useRecoilValue } from 'recoil';
-import client from '../config/api.config';
-import userAtom from '../atoms/userAtom';
+import client from '../../config/api.config';
+import userAtom from '../../atoms/userAtom';
 
 /**
  * 팔로잉 리스트를 가져옵니다.
