+++ src/api/get/getMyInfo.api.js
@@ -1,6 +1,6 @@
 import { useRecoilValue } from 'recoil';
-import client from '../config/api.config';
-import userAtom from '../atoms/userAtom';
+import client from '../../config/api.config';
+import userAtom from '../../atoms/userAtom';
 
 //프로필 수정 - 원래 있던 정보를 불러와야 할 때 사용합니다!
 const getMyInfo = async () => {
