+++ src/api/get/getProductDetail.api.js
@@ -1,26 +1,25 @@
 import { useState } from 'react';
-import client from '../config/api.config';
+import client from '../../config/api.config';
 
-const getPostDetail = () => {
+const getProductDetail = () => {
   const [loading, setLoading] = useState(false);
   const [error, setError] = useState(null);
 
-  const fetchPost = async (id) => {
+  const fetchProduct = async (id) => {
     setLoading(true);
     const token = JSON.parse(localStorage.getItem('user')).token;
 
     try {
-      const response = await client.get(`/post/${id}`, {}, client.BothType(token));
-      return response.data.post;
+      const response = await client.get(`/product/detail/${id}`, {}, client.BothType(token));
+      return response.data.product;
     } catch (error) {
-      console.error(error);
       setError(error);
       return;
     }
     setLoading(false);
   };
 
-  return { fetchPost, loading, error };
+  return { fetchProduct, loading, error };
 };
 
-export default getPostDetail;
+export default getProductDetail;
