+++ src/api/get/getSearchUser.api.js
@@ -1,4 +1,6 @@
-import client from '../config/api.config';
+import { useRecoilValue } from 'recoil';
+import client from '../../config/api.config';
+import userAtom from '../../atoms/userAtom';
 
 /**
  * accountname, username을 통해서 유저를 검색합니다.
