+++ src/api/post/postComments.api.js
@@ -1,5 +1,5 @@
 import { useState } from 'react';
-import client from '../config/api.config';
+import client from '../../config/api.config';
 
 const postComments = () => {
   const [loading, setLoading] = useState(false);
