+++ src/components/feed/Comment.jsx
@@ -4,8 +4,8 @@ import verticalIcon from '../../assets/img/icon-more-vertical.svg';
 import useFormatDate from '../../hooks/useFormatDate';
 import { useRecoilValue, useSetRecoilState } from 'recoil';
 import { bottomSheetStateAtom, bottomSheetOptions } from '../../atoms/bottomSheetStateAtom';
-import deleteCommentsQuery from '../../api/deleteComments.api';
-import postCommentsReportQuery from '../../api/postCommentsReport.api';
+import deleteCommentsQuery from '../../api/delete/deleteComments.api';
+import postCommentsReportQuery from '../../api/post/postCommentsReport.api';
 import { useMutation } from '@tanstack/react-query';
 import userAtom from '../../atoms/userAtom';
 import useBottomSheetOptions from '../../hooks/useBottomSheetOptions';
