+++ src/components/feed/PostCard.jsx
@@ -4,7 +4,7 @@ import heartFilled from '../../assets/img/heart-filled.svg';
 import comment from '../../assets/img/message-circle.svg';
 import { useNavigate } from 'react-router-dom';
 import pageUlrConfig from '../../config/pageUrlConfig';
-import postHeart from '../../api/postHeart.api';
+import postHeart from '../../api/post/postHeart.api';
 import { useState } from 'react';
 import useFormatDate from '../../hooks/useFormatDate';
 import { IconLabelBtn } from '../Buttons';
